move sidebar items and methods to the correct page object
